pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "docex"
version = "0.2.0"
description = "A robust, lightweight, and developer-friendly document management and transport system for Python."
readme = "README.md"
license = { file = "LICENSE" }
authors = [
    { name = "Tommy Jiang", email = "tommyscosc@scos.ai" }
]
dependencies = [
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "pyyaml>=6.0.0",
    "pytest>=7.0.0",
    "boto3>=1.26.0",
    "pydantic>=2.0.0",
    "paramiko>=3.4.0",
    "aiohttp>=3.9.0",
    "python-dotenv>=1.0.0",
    "click>=8.0.0"
]
requires-python = ">=3.11"

[project.scripts]
docex = "docflow.cli:cli" 